refactor Subscriptions Container
current subscription refactor


subscription history refactor


subscription history refactor


refactor premium credits sections


refactor refund policy
