+++ client/app/bundles/HelloWorld/containers/Subscriptions.jsx
@@ -6,7 +6,10 @@ import pluralize from 'pluralize';
 import SubscriptionStatus from '../components/subscriptions/subscription_status';
 import PaymentModal from '../components/subscriptions/select_credit_card_modal';
 import CurrentSubscription from '../components/subscriptions/current_subscription';
+import SubscriptionHistory from '../components/subscriptions/subscription_history';
 import PremiumConfirmationModal from '../components/subscriptions/premium_confirmation_modal';
+import RefundPolicy from '../components/subscriptions/refund_policy';
+import PremiumCredits from '../components/subscriptions/premium_credits';
 import getAuthToken from '../components/modules/get_auth_token';
 
 export default class extends React.Component {
@@ -34,83 +37,6 @@ export default class extends React.Component {
 
   updateSubscriptionStatus(subscription) {
     this.setState({ subscriptionStatus: subscription, showPremiumConfirmationModal: true, showPurchaseModal: false, });
-  }
-
-  subscriptionHistoryRows() {
-    const rows = [];
-    this.state.subscriptions.forEach((sub) => {
-      const startD = moment(sub.start_date);
-      const endD = moment(sub.expiration);
-      const duration = endD.diff(startD, 'months');
-      const matchingTransaction = this.props.premiumCredits.find(transaction => (transaction.source_id === sub.id && transaction.source_type === 'Subscription' && transaction.amount > 0));
-      if (matchingTransaction) {
-        const amountCredited = matchingTransaction.amount > 6
-          ? Math.round(matchingTransaction.amount / 7)
-          : 1;
-        rows.push(
-          <tr key={`${matchingTransaction.id}-credit-subscription-table`} className="subscription-row text-center">
-            <td colSpan="5">
-              Your school purchased School Premium during your subscription, so we credited your account with {`${amountCredited} ${pluralize('week', amountCredited)}`} of Teacher Premium.
-            </td>
-          </tr>
-        );
-      }
-      rows.push(
-        <tr key={`${sub.id}-subscription-table`}>
-          <td>{moment(sub.created_at).format('MMMM Do, YYYY')}</td>
-          <td>{sub.account_type}</td>
-          <td>{this.paymentContent(sub)}</td>
-          <td>{`${duration} ${pluralize('month', duration)}`}</td>
-          <td>{`${startD.format('MM/DD/YY')} - ${endD.format('MM/DD/YY')}`}</td>
-        </tr>
-      );
-    });
-    return rows;
-  }
-
-  paymentContent(subscription) {
-    const currentUserId = document.getElementById('current-user-id').getAttribute('content');
-    if (subscription.contact_user_id === Number(currentUserId)) {
-      if (subscription.payment_amount) {
-        return `$${subscription.payment_amount / 100}`;
-      } else if (subscription.payment_method === 'Premium Credit') {
-        return subscription.payment_method;
-      }
-    }
-    return '--';
-  }
-
-  subscriptionHistory() {
-    const subscriptionHistoryRows = this.subscriptionHistoryRows();
-    let content;
-    if (subscriptionHistoryRows.length > 0) {
-      content = (
-        <table>
-          <tbody>
-            <tr>
-              <th>Purchase Date</th>
-              <th>Subscription</th>
-              <th>Payment</th>
-              <th>Length</th>
-              <th>Start & End Date</th>
-            </tr>
-            {subscriptionHistoryRows}
-          </tbody>
-        </table>);
-    } else {
-      content = (
-        <div className="empty-state flex-row justify-content">
-          <h3>You have not yet started a Quill Premium Subscription</h3>
-          <span>Purchase Quill Premium or apply credits to get access to Premium reports.</span>
-        </div>
-      );
-    }
-    return (
-      <section className="subscription-history">
-        <h2>Premium Subscription History</h2>
-        {content}
-      </section>
-    );
   }
 
   availableAndEarnedCredits() {
@@ -150,34 +76,6 @@ export default class extends React.Component {
     return <button type="button" id="purchase-btn" data-toggle="modal" onClick={this.purchasePremiu} className="q-button button cta-button bg-orange text-white">Update Card</button>;
   }
 
-  premiumCreditsTable() {
-    const creditRows = this.props.premiumCredits.map((credit) => {
-      // if it is less than one week, we round up to 1
-      const amountCredited = credit.amount > 6
-        ? Math.round(credit.amount / 7)
-        : 1;
-      return (
-        <tr key={`credit-${credit.id}-premium-credit-table`}>
-          <td className="date-received">{moment(credit.created_at).format('MMMM Do, YYYY')}</td>
-          <td className="amount-credited">{`${amountCredited} ${pluralize('week', amountCredited)}`}</td>
-          <td className="action">{credit.action || 'Lorem ipsum dolor sit amet, consectetur adipisicing elit!'}</td>
-        </tr>
-      );
-    });
-    return (
-      <table className="premium-credits-table">
-        <tbody>
-          <tr>
-            <th>Date Received</th>
-            <th>Amount Credited</th>
-            <th>Action</th>
-          </tr>
-          {creditRows}
-        </tbody>
-      </table>
-    );
-  }
-
   currentSubscription(newSub) {
     if (!this.state.subscriptionStatus || this.state.subscriptionStatus.expired) {
       return newSub;
@@ -203,43 +101,6 @@ export default class extends React.Component {
         });
       }
     });
-  }
-
-  availableCredits() {
-    let button;
-    if (this.state.availableCredits > 0) {
-      button = <button onClick={this.redeemPremiumCredits} className="q-button cta-button bg-orange has-credit">Redeem Premium Credits</button>;
-    } else {
-      button = <a href="/" className="q-button button cta-button bg-orange">Earn Premium Credits</a>;
-    }
-    const monthsOfCredit = Math.round((this.state.availableCredits / 30.42) * 10) / 10;
-    const whiteIfNoCredit = monthsOfCredit === 0 ? 'no-credits' : null;
-    return (
-      <div className={`${whiteIfNoCredit} available-credit flex-row vertically-centered space-between`}>
-        <div className="credit-quantity">
-          You have <span>{`${monthsOfCredit} ${pluralize('month', monthsOfCredit)} `}</span> of Teacher Premium Credit available.
-        </div>
-        {button}
-      </div>
-    );
-  }
-
-  premiumCredits() {
-    if (!this.props.premiumCredits || this.props.premiumCredits < 1) {
-      return this.availableCredits();
-    }
-    const monthsOfCredit = Math.round(((this.state.earnedCredits / 30.42) * 10) / 10);
-    return (
-      <section>
-        <div className="flex-row space-between">
-          <h2>Quill Teacher Premium Credits</h2>
-          <a className="green-link" href="">How to earn more Premium credit</a>
-        </div>
-        {this.availableCredits()}
-        {this.premiumCreditsTable()}
-        <span className="total-premium-credits"><span className="total-header">Total Premium Credits Earned:</span> {`${monthsOfCredit} ${pluralize('month', monthsOfCredit)}`}</span>
-      </section>
-    );
   }
 
   updateCard() {
@@ -271,14 +132,17 @@ export default class extends React.Component {
           subscriptionStatus={this.state.subscriptionStatus}
           lastFour={this.props.lastFour}
         />
-        {this.subscriptionHistory()}
-        {this.premiumCredits()}
-        <section className="refund-policy">
-          <h2>Refund Policy</h2>
-          <p>
-            If you purchase a Teacher Premium subscription, and then your school purchases a School Premium subscription, you will be refunded the remainder of your Teacher Premium as Quill Premium Credit. You can redeem your Premium Credit anytime you do not currently have an active subscription, and you will be resubscribed to Quill Premium for the amount of time you have in credit. If you would like to receive a full refund there is a grace period of 5 days from the day of the renewal.
-          </p>
-        </section>
+        <SubscriptionHistory
+          subscriptions={this.state.subscriptions}
+          premiumCredits={this.props.premiumCredits}
+        />
+        <PremiumCredits
+          earnedCredits={this.state.earnedCredits}
+          availableCredits={this.state.availableCredits}
+          premiumCredits={this.props.premiumCredits}
+          redeemPremiumCredits={this.redeemPremiumCredits}
+        />
+        <RefundPolicy />
         <PremiumConfirmationModal show={this.state.showPremiumConfirmationModal} hideModal={this.hidePremiumConfirmationModal} subscription={this.state.subscriptionStatus} />
         <PaymentModal show={this.state.showPaymentModal} hideModal={this.hidePaymentModal} lastFour={this.props.lastFour} updateSubscriptionStatus={this.updateSubscriptionStatus} />
